[python/modules/menu_creation.py] Create module
This module is used to house all the functions used to display an options menu and execute its options, all of which must be defined in a dictionary.

All the menu-printing and option-executing logic is included with this commit, but I have yet to add type and format checks to validate the correctness of the dictionary provided as input.