[python/{genscript_tools,modules}] Add a tags system
This reworks the mechanism I had introduced two commits ago to use tags to tag the functions that have to be executed with root access.

Now, the tag is much more different, being called "#ROOT", and there is also a tag that allows to specify which commands need to have user input appended to them.

The "pipe" tag used to mark functions that need to be run with pipe_commands() has been renamed to "#PIPE".