[python/modules] Get rid of the hardcoded usage of doas to run software as root
This is solved by adding a get_privilege_escalation_command() function to the program_tools.py module, which will poll the system to check if any of the programs used for privilege elevation defined in the PRIVILEGE_ELEVATION_COMMANDS list is being used.

If none of the programs are present, the function will return "pkexec" as a result, which should be available on almost any Linux installation because of being provided by polkit.